Add pkg/auth/init.go
Pull out all the code for initialising the auth server to try and keep
the files a little more focused
