+++ pkg/server/auth/auth_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/oauth2-proxy/mockoidc"
 	"github.com/onsi/gomega"
 	. "github.com/onsi/gomega"
-	"github.com/weaveworks/weave-gitops/pkg/featureflags"
 	"github.com/weaveworks/weave-gitops/pkg/server/auth"
 	"golang.org/x/crypto/bcrypt"
 	corev1 "k8s.io/api/core/v1"
@@ -188,140 +187,4 @@ func TestRateLimit(t *testing.T) {
 	res3, err := http.Post(s.URL+"/oauth2/sign_in", "application/json", bytes.NewReader([]byte(`{"password":"my-secret-password"}`)))
 	g.Expect(err).NotTo(HaveOccurred())
 	g.Expect(res3).To(HaveHTTPStatus(http.StatusOK))
-}
-
-func TestInitAuthServer(t *testing.T) {
-	g := gomega.NewGomegaWithT(t)
-
-	m, err := mockoidc.Run()
-	g.Expect(err).NotTo(HaveOccurred())
-
-	t.Cleanup(func() {
-		_ = m.Shutdown()
-	})
-
-	initTests := []struct{
-		name string
-		secrets []*corev1.Secret
-		cliOIDCConfig auth.OIDCConfig
-		oidcSecretName string
-		expectErr bool
-		clusterUserFlag string
-		oidcEnabledFlag string
-	}{
-		{
-			name: "basic test",
-			secrets: []*corev1.Secret{
-				makeOIDCSecret(m.Config(), auth.DefaultOIDCAuthSecretName),
-				makeClusterUserSecret("my-secret-password", auth.ClusterUserAuthSecretName),
-			},
-			cliOIDCConfig: auth.OIDCConfig{},
-			oidcSecretName: auth.DefaultOIDCAuthSecretName,
-			expectErr: false,
-			clusterUserFlag: "true",
-			oidcEnabledFlag: "true",
-		},
-		{
-			name: "OIDC Only",
-			secrets: []*corev1.Secret{
-				makeOIDCSecret(m.Config(), auth.DefaultOIDCAuthSecretName),
-			},
-			cliOIDCConfig: auth.OIDCConfig{},
-			oidcSecretName: auth.DefaultOIDCAuthSecretName,
-			expectErr: false,
-			clusterUserFlag: "false",
-			oidcEnabledFlag: "true",
-		},
-		{
-			name: "OIDC alt-secret",
-			secrets: []*corev1.Secret{
-				makeOIDCSecret(m.Config(), "alternate-oidc-secret"),
-			},
-			cliOIDCConfig: auth.OIDCConfig{},
-			oidcSecretName: "alternate-oidc-secret",
-			expectErr: false,
-			clusterUserFlag: "false",
-			oidcEnabledFlag: "true",
-		},
-		{
-			name: "OIDC via CLI",
-			secrets: []*corev1.Secret{},
-			cliOIDCConfig: auth.OIDCConfig{
-				IssuerURL: m.Config().Issuer,
-				ClientID: m.Config().ClientID,
-				ClientSecret: m.Config().ClientSecret,
-				RedirectURL: "example.invalid/oauth2/callback",
-			},
-			oidcSecretName: auth.DefaultOIDCAuthSecretName,
-			expectErr: false,
-			clusterUserFlag: "false",
-			oidcEnabledFlag: "true",
-		},
-		{
-			name: "User only",
-			secrets: []*corev1.Secret{
-				makeClusterUserSecret("my-secret-password", auth.ClusterUserAuthSecretName),
-			},
-			cliOIDCConfig: auth.OIDCConfig{},
-			oidcSecretName: auth.DefaultOIDCAuthSecretName,
-			expectErr: false,
-			clusterUserFlag: "true",
-			oidcEnabledFlag: "false",
-		},
-	}
-
-	for _, tt := range initTests {
-		t.Run( tt.name, func(t *testing.T){
-			partialKubernetesClient := ctrlclient.NewClientBuilder()
-
-			// This is because I can't (be bothered to) figure out how to []*secret -> []*client.Object
-			for _, obj := range tt.secrets {
-				partialKubernetesClient.WithObjects(obj)
-			}
-
-			fakeKubernetesClient := partialKubernetesClient.Build()
-
-			srv, err := auth.InitAuthServer(context.Background(), logr.Discard(), fakeKubernetesClient, tt.cliOIDCConfig, tt.oidcSecretName, "", false)
-
-			if tt.expectErr {
-				g.Expect(err).To(HaveOccurred())
-				g.Expect(srv).To(BeNil())
-			} else {
-				g.Expect(err).NotTo(HaveOccurred())
-				g.Expect(srv).NotTo(BeNil())
-			}
-
-			// Can check (somewhat) what's happened by making sure flags are set
-			g.Expect(featureflags.Get(auth.FeatureFlagClusterUser)).To(Equal(tt.clusterUserFlag))
-			g.Expect(featureflags.Get(auth.FeatureFlagOIDCAuth)).To(Equal(tt.oidcEnabledFlag))
-		})
-	}
-}
-
-func makeOIDCSecret(oidcConfig *mockoidc.Config, secretName string) *corev1.Secret {
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: secretName,
-			Namespace: "flux-system",
-		},
-		Data: map[string][]byte{
-			"issuerURL": []byte(oidcConfig.Issuer),
-			"clientID": []byte(oidcConfig.ClientID),
-			"clientSecret": []byte(oidcConfig.ClientSecret),
-			"redirectURL": []byte("test.invalid/oauth2/callback"),
-		},
-	}
-}
-
-func makeClusterUserSecret(password, secretName string) *corev1.Secret {
-	hashed, _ := bcrypt.GenerateFromPassword([]byte(password), bcrypt.DefaultCost)
-	return &corev1.Secret{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: secretName,
-			Namespace: "flux-system",
-		},
-		Data: map[string][]byte{
-			"password": hashed,
-		},
-	}
 }
